+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/index.tsx","./src/main.tsx","./src/app/libs/axiosinstance.old.ts","./src/app/libs/axiosinstance.ts","./src/app/providers/query-client/config.ts","./src/app/providers/query-client/index.ts","./src/app/providers/query-client/provider.tsx","./src/entities/comment/api/index.ts","./src/entities/comment/model/index.ts","./src/entities/comment/model/request.types.ts","./src/entities/comment/model/response.types.ts","./src/entities/comment/queries/index.ts","./src/entities/comment/store/usecommentstore.ts","./src/entities/comment/ui/index.ts","./src/entities/comment/ui/comment-item/index.tsx","./src/entities/post/api/index.ts","./src/entities/post/model/index.ts","./src/entities/post/model/request.typee.ts","./src/entities/post/model/response.types.ts","./src/entities/post/queries/index.ts","./src/entities/post/store/usepoststore.ts","./src/entities/tag/api/index.ts","./src/entities/tag/model/index.ts","./src/entities/tag/model/types.ts","./src/entities/tag/queries/index.ts","./src/entities/user/api/index.ts","./src/entities/user/model/index.ts","./src/entities/user/model/types.ts","./src/entities/user/queries/index.ts","./src/entities/user/store/useuserstore.ts","./src/features/comment/mutations/index.ts","./src/features/comment/ui/create-comment-form/index.tsx","./src/features/comment/ui/edit-comment-form/index.tsx","./src/features/pagination/ui/pagination-control/index.tsx","./src/features/post/mutations/index.ts","./src/features/post/ui/create-post-form/index.tsx","./src/features/post/ui/edit-post-form/index.tsx","./src/features/post/ui/post-detail/index.tsx","./src/features/post/ui/post-search/index.tsx","./src/features/post/ui/sort-order-selector/index.tsx","./src/features/post/ui/sort-selector/index.tsx","./src/features/post/ui/tag-filter/index.tsx","./src/features/user/ui/user-profile/index.tsx","./src/pages/postsmanagerpage.tsx","./src/shared/model/index.ts","./src/shared/model/types/api.ts","./src/shared/model/types/user.ts","./src/shared/ui/index.ts","./src/shared/ui/boundaries/fetch-suspense/fetchsuspense.tsx","./src/shared/ui/button/index.tsx","./src/shared/ui/card/index.tsx","./src/shared/ui/dialog/index.tsx","./src/shared/ui/input/index.tsx","./src/shared/ui/select/index.tsx","./src/shared/ui/table/index.tsx","./src/shared/ui/text-highlighter/index.tsx","./src/shared/ui/textarea/index.tsx","./src/widgets/comment/ui/index.ts","./src/widgets/comment/ui/comment-list/index.tsx","./src/widgets/layout/ui/index.ts","./src/widgets/layout/ui/default-layout/index.tsx","./src/widgets/layout/ui/footer/index.tsx","./src/widgets/layout/ui/header/index.tsx","./src/widgets/post/ui/post-table-container/index.tsx","./vite-env.d.ts"],"version":"5.9.2"}
+{"root":["./src/app.tsx","./src/index.tsx","./src/main.tsx","./src/app/libs/axiosinstance.old.ts","./src/app/libs/axiosinstance.ts","./src/app/providers/query-client/config.ts","./src/app/providers/query-client/index.ts","./src/app/providers/query-client/provider.tsx","./src/entities/comment/api/index.ts","./src/entities/comment/model/index.ts","./src/entities/comment/model/request.types.ts","./src/entities/comment/model/response.types.ts","./src/entities/comment/queries/index.ts","./src/entities/comment/store/usecommentstore.ts","./src/entities/comment/ui/index.ts","./src/entities/comment/ui/comment-item/index.tsx","./src/entities/post/api/index.ts","./src/entities/post/model/index.ts","./src/entities/post/model/request.typee.ts","./src/entities/post/model/response.types.ts","./src/entities/post/queries/index.ts","./src/entities/post/store/usepoststore.ts","./src/entities/tag/api/index.ts","./src/entities/tag/model/index.ts","./src/entities/tag/model/types.ts","./src/entities/tag/queries/index.ts","./src/entities/user/api/index.ts","./src/entities/user/model/index.ts","./src/entities/user/model/types.ts","./src/entities/user/queries/index.ts","./src/entities/user/store/useuserstore.ts","./src/features/comment/mutations/index.ts","./src/features/comment/ui/create-comment-form/index.tsx","./src/features/comment/ui/edit-comment-form/index.tsx","./src/features/pagination/ui/pagination-control/index.tsx","./src/features/post/hooks/usepostfilters.ts","./src/features/post/mutations/index.ts","./src/features/post/ui/create-post-form/index.tsx","./src/features/post/ui/edit-post-form/index.tsx","./src/features/post/ui/post-detail/index.tsx","./src/features/post/ui/post-search/index.tsx","./src/features/post/ui/sort-order-selector/index.tsx","./src/features/post/ui/sort-selector/index.tsx","./src/features/post/ui/tag-filter/index.tsx","./src/features/user/ui/user-profile/index.tsx","./src/pages/postsmanagerpage.tsx","./src/shared/model/index.ts","./src/shared/model/types/api.ts","./src/shared/model/types/user.ts","./src/shared/ui/index.ts","./src/shared/ui/boundaries/fetch-suspense/fetchsuspense.tsx","./src/shared/ui/button/index.tsx","./src/shared/ui/card/index.tsx","./src/shared/ui/dialog/index.tsx","./src/shared/ui/input/index.tsx","./src/shared/ui/select/index.tsx","./src/shared/ui/table/index.tsx","./src/shared/ui/text-highlighter/index.tsx","./src/shared/ui/textarea/index.tsx","./src/widgets/comment/ui/index.ts","./src/widgets/comment/ui/comment-list/index.tsx","./src/widgets/layout/ui/index.ts","./src/widgets/layout/ui/default-layout/index.tsx","./src/widgets/layout/ui/footer/index.tsx","./src/widgets/layout/ui/header/index.tsx","./src/widgets/post/ui/post-table-container/index.tsx","./vite-env.d.ts"],"version":"5.9.2"}
